Add files via upload
First version.
In the future I would like to add more options via flags. (For example, to enable verbose using the -v flag. Currently verbose would need to be enabled manually by changing the options the script uses for tar and rsync commands) (Another future addition may be to allow the user to specify (in a textfile) which directories and files he would like to be excluded from the tar archives and which directories he would like to be synced via rsync. Currently this needs to be changed manually in the script.)
I would also like to add a feature where disk space is calculated at the beginning of the script. If the directories to be backed up are found to exceed the disk space, the script will: Print out a warning with a summary of how much disk space is left and how much disk space the directories to be backed up would take. Then the script will terminate.